Hand wrote a sprite spec file--almost died doing it.
Added ships.tiff which is a vertical AND horizontal mirror of ships.gif
because the sprite shader need more knobs to control how it indexes into
the sprite image.

Notice that the sprite shader math is odd and in low res images can have
rounding error and mis-index into the sprite sheet.

Altered the game window to be the aspect ratio we need of 240w x 320h.