Began adding a visualization similar to jstest-gtk.
There is an additional scne you can run called

htdf-calibrate-gamepad

which is only the visualization of the gamepad.

I'll add more buttons and features to it in a bit, but at least the
two analog sticks are visualized currently. When we get some text
components, or I make a pile of textures, then I can make it prettier.